Move version.py to be more correct.
If it's in the package there can be issues with importing if deps are
not installed
