+++ MANIFEST.in
@@ -1,2 +1,3 @@
 include README.md
 include routemaster/config/schema.yaml
+include version.py
