+++ setup.py
@@ -1,8 +1,7 @@
 """Package setup."""
 
+import version
 from setuptools import setup, find_packages
-
-from routemaster import version
 
 with open('README.md', 'r', encoding='utf-8') as f:
     long_description = f.read()
